// src/app/page/master/map-pegawai-produkdiklat/map-pegawai-produkdiklat.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapPegawaiProdukdiklat {
	kdProduk?: string;
	kdEvent?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}